Flows: Bin, BinSku, Channel, Listing Perform tasks: Remove unused functions, add checks in dto helper, service class constraints(except listing flow), fixed( converted int to Long, removed rollback exception from readOnly), add AbstractPojo to include create, update date and version

+++ assure/src/main/java/com/increff/assure/dao/ChannelDao.java
@@ -24,7 +24,7 @@ public class ChannelDao extends AbstractDao {
         em.persist(p);
     }
 
-    public ChannelPojo select(int id) {
+    public ChannelPojo select(Long id) {
         TypedQuery<ChannelPojo> query = getQuery(SELECT_ID, ChannelPojo.class);
         query.setParameter("id", id);
         return getSingle(query);
